+++ CHANGELOG.md
@@ -200,6 +200,8 @@ All notable changes to this project will be documented in this file.
   - Updated the validation script to replicate the newly added filter from the cleaning process, ensuring row count validations now pass.
   - Introduced a new function that creates boolean flag columns for complex-type columns and removes the original complex columns from the source DataFrame. These flag columns are now used in PointBlank validations to check for True values.
 
+- Fixed add_previous_value_column to return the value/null from the previous import_date, not the last known value from any point in time.
+
 ### Improved
 - Moved postcode corrections dictionary into a csv file in s3.